Harden KeysetContinuationToken decode and encode
The token crosses a trust boundary: it is base64 in a client-supplied URL, so
every coordinate has to be validated on the way in rather than trusted because
the server minted the original.

Decode now refuses an absent or non-positive resource-type id, an absent or
non-positive surrogate id, more values than the 3-key sort cap allows, a null
element inside the value array, and an absent or out-of-range sort phase.
Previously each of these produced a KeysetPosition that looked well-formed and
failed later -- as a malformed seek predicate, or worse, as a silently wrong
page. The unreachable DecoderFallbackException filter is dropped.

Encode rejects what it cannot round-trip: null elements, and strings holding an
unpaired surrogate, which UTF-8 encoding would replace with U+FFFD, handing the
caller a token that decodes to a different boundary than the one it was given.

BoundaryResourceTypeId narrows int -> short, matching every other resource-type
id in the assembly, so an out-of-range value now fails JSON deserialisation
before any validation runs.

A golden wire-format vector pins the encoding in both directions, so a field
rename cannot silently invalidate every token already in flight.

Copilot-Session: 75061886-abb9-46f7-89dd-a1e40f3d4fff

+++ src/Core/Ignixa.Search.Sql/Ast/KeysetPosition.cs
@@ -9,16 +9,18 @@ namespace Ignixa.Search.Sql.Ast;
 /// <param name="BoundaryValues">
 /// One value per active sort key for <paramref name="Phase"/>: <c>SortSpec.Keys.Count</c> under
 /// <see cref="SortPhase.Valued"/>, one fewer under <see cref="SortPhase.MissingPrimary"/>, since the primary
-/// key contributes no value when it is absent. Empty on an unsorted query.
+/// key contributes no value when it is absent. Empty on an unsorted query, and never more than the three keys
+/// a <c>_sort</c> may carry.
 /// </param>
 /// <param name="BoundaryResourceTypeId">
-/// The type of the boundary row. Always carried; a token-to-<see cref="PageSpec"/> adapter discards it for a
-/// custom sort, where <see cref="PageSpec"/> requires a typeless boundary.
+/// The type of the boundary row, typed to match the schema's <c>smallint</c>. Always carried; a
+/// token-to-<see cref="PageSpec"/> adapter discards it for a custom sort, where <see cref="PageSpec"/>
+/// requires a typeless boundary.
 /// </param>
 /// <param name="BoundarySurrogateId">The boundary row's surrogate id, which breaks the final tie.</param>
 /// <param name="Phase">The segment this position was reached in. A position never crosses a phase boundary.</param>
 public sealed record KeysetPosition(
     IReadOnlyList<string> BoundaryValues,
-    int BoundaryResourceTypeId,
+    short BoundaryResourceTypeId,
     long BoundarySurrogateId,
     SortPhase Phase);
